// FrontEnd/src/models/BeltRequirements/BeltRequirements.ts
export class BeltRequirements {
  beltOrder: number = 1;
  beltRank: string = 'White Belt';
  forms: string[] = [];
  stances: string[] = [];
  blocks: string[] = [];
  punches: string[] = [];
  kicks: string[] = [];
  jumps: string[] = [];
  falling: string[] = [];
  oneSteps: string[] = [];
  selfDefense: string[] = [];
  comments?: string;
  color?: string;
  textColor?: string;
}